+++ examples/computeinstanceagent/instance_agent_plugins/instance_agent_plugins.tf
@@ -1,0 +1,183 @@
+// Copyright (c) 2017, 2020, Oracle and/or its affiliates. All rights reserved.
+// Licensed under the Mozilla Public License v2.0
+
+variable "tenancy_ocid" {
+
+}
+
+variable "user_ocid" {
+
+}
+
+variable "fingerprint" {
+
+}
+
+variable "private_key_path" {
+
+}
+
+variable "region" {
+
+}
+
+variable "compartment_ocid" {
+
+}
+
+provider "oci" {
+  tenancy_ocid     = var.tenancy_ocid
+  user_ocid        = var.user_ocid
+  fingerprint      = var.fingerprint
+  private_key_path = var.private_key_path
+  region           = var.region
+}
+
+variable "num_instances" {
+  default = "1"
+}
+
+variable "instance_shape" {
+  default = "VM.Standard.E3.Flex"
+}
+
+variable "instance_ocpus" {
+  default = 1
+}
+
+variable "instance_shape_config_memory_in_gbs" {
+  default = 1
+}
+
+variable "instance_image_ocid" {
+  type = map(string)
+
+  default = {
+    # See https://docs.us-phoenix-1.oraclecloud.com/images/
+    # Oracle-provided image "Oracle-Linux-7.5-2018.10.16-0"
+    us-phoenix-1   = "ocid1.image.oc1.phx.aaaaaaaaoqj42sokaoh42l76wsyhn3k2beuntrh5maj3gmgmzeyr55zzrwwa"
+    us-ashburn-1   = "ocid1.image.oc1.iad.aaaaaaaageeenzyuxgia726xur4ztaoxbxyjlxogdhreu3ngfj2gji3bayda"
+    eu-frankfurt-1 = "ocid1.image.oc1.eu-frankfurt-1.aaaaaaaaitzn6tdyjer7jl34h2ujz74jwy5nkbukbh55ekp6oyzwrtfa4zma"
+    uk-london-1    = "ocid1.image.oc1.uk-london-1.aaaaaaaa32voyikkkzfxyo4xbdmadc2dmvorfxxgdhpnk6dw64fa3l4jh7wa"
+  }
+}
+
+variable "flex_instance_image_ocid" {
+  type = map(string)
+  default = {
+    us-phoenix-1 = "ocid1.image.oc1.phx.aaaaaaaa6hooptnlbfwr5lwemqjbu3uqidntrlhnt45yihfj222zahe7p3wq"
+    us-ashburn-1 = "ocid1.image.oc1.iad.aaaaaaaa6tp7lhyrcokdtf7vrbmxyp2pctgg4uxvt4jz4vc47qoc2ec4anha"
+    eu-frankfurt-1 = "ocid1.image.oc1.eu-frankfurt-1.aaaaaaaadvi77prh3vjijhwe5xbd6kjg3n5ndxjcpod6om6qaiqeu3csof7a"
+    uk-london-1 = "ocid1.image.oc1.uk-london-1.aaaaaaaaw5gvriwzjhzt2tnylrfnpanz5ndztyrv3zpwhlzxdbkqsjfkwxaq"
+  }
+}
+
+variable "db_size" {
+  default = "50" # size in GBs
+}
+
+
+resource "oci_core_instance" "test_instance" {
+  count               = var.num_instances
+  availability_domain = data.oci_identity_availability_domain.ad.name
+  compartment_id      = var.compartment_ocid
+  display_name        = "TestInstance${count.index}"
+  shape               = var.instance_shape
+
+  shape_config {
+    ocpus = "${var.instance_ocpus}"
+    memory_in_gbs = "${var.instance_shape_config_memory_in_gbs}"
+  }
+
+  create_vnic_details {
+    subnet_id        = oci_core_subnet.test_subnet.id
+    display_name     = "Primaryvnic"
+    assign_public_ip = true
+    hostname_label   = "tfexampleinstance${count.index}"
+  }
+
+  source_details {
+    source_type = "image"
+    source_id = var.flex_instance_image_ocid[var.region]
+  }
+
+  agent_config {
+    are_all_plugins_disabled = false
+    is_management_disabled = true
+    is_monitoring_disabled = true
+    plugins_config {
+        name = "Compute Instance Monitoring"
+        desired_state = "ENABLED"
+      }
+  }
+}
+
+resource "oci_core_vcn" "test_vcn" {
+  cidr_block     = "10.1.0.0/16"
+  compartment_id = var.compartment_ocid
+  display_name   = "TestVcn"
+  dns_label      = "testvcn"
+}
+
+resource "oci_core_internet_gateway" "test_internet_gateway" {
+  compartment_id = var.compartment_ocid
+  display_name   = "TestInternetGateway"
+  vcn_id         = oci_core_vcn.test_vcn.id
+}
+
+resource "oci_core_default_route_table" "default_route_table" {
+  manage_default_resource_id = oci_core_vcn.test_vcn.default_route_table_id
+  display_name               = "DefaultRouteTable"
+
+  route_rules {
+    destination       = "0.0.0.0/0"
+    destination_type  = "CIDR_BLOCK"
+    network_entity_id = oci_core_internet_gateway.test_internet_gateway.id
+  }
+}
+
+resource "oci_core_subnet" "test_subnet" {
+  availability_domain = data.oci_identity_availability_domain.ad.name
+  cidr_block          = "10.1.20.0/24"
+  display_name        = "TestSubnet"
+  dns_label           = "testsubnet"
+  security_list_ids   = [oci_core_vcn.test_vcn.default_security_list_id]
+  compartment_id      = var.compartment_ocid
+  vcn_id              = oci_core_vcn.test_vcn.id
+  route_table_id      = oci_core_vcn.test_vcn.default_route_table_id
+  dhcp_options_id     = oci_core_vcn.test_vcn.default_dhcp_options_id
+}
+
+data "oci_identity_availability_domain" "ad" {
+  compartment_id = var.tenancy_ocid
+  ad_number      = 1
+}
+
+data "oci_computeinstanceagent_instance_agent_plugins" "test_instance_agent_plugins" {
+  compartment_id = var.compartment_ocid
+  instanceagent_id = oci_core_instance.test_instance.0.id
+}
+
+output "agent_plugins" {
+  value = [data.oci_computeinstanceagent_instance_agent_plugins.test_instance_agent_plugins]
+}
+
+variable "instance_available_plugin_os_name" {
+  default = "Oracle Linux"
+}
+
+variable "instance_available_plugin_os_version" {
+  default = "7.8"
+}
+
+
+data "oci_computeinstanceagent_instance_available_plugins" "test_instance_available_plugins" {
+  #Required
+  compartment_id = var.compartment_ocid
+  os_name    = var.instance_available_plugin_os_name
+  os_version = var.instance_available_plugin_os_version
+}
+
+output "available_plugins" {
+  value = [data.oci_computeinstanceagent_instance_available_plugins.test_instance_available_plugins]
+}